+++ lib/response.js
@@ -619,7 +619,7 @@ internals.Response.Payload = class extends Stream.Readable {
         this._encoding = options.encoding;
     }
 
-    _read(/* size */) {
+    _read(size) {
 
         if (this._prefix) {
             this.push(this._prefix, this._encoding);
