+++ lib/route.js
@@ -65,7 +65,7 @@ exports = module.exports = internals.Route = class {
         const handlerDefaults = Handler.defaults(method, handler, core);
         let base = Hoek.applyToDefaultsWithShallow(core.settings.routes, handlerDefaults, ['bind']);
         base = Hoek.applyToDefaultsWithShallow(base, realm.settings, ['bind']);
-        this.settings = Hoek.applyToDefaultsWithShallow(base, config || {}, ['bind', 'validate.headers', 'validate.payload', 'validate.params', 'validate.query']);
+        this.settings = Hoek.applyToDefaultsWithShallow(base, config, ['bind', 'validate.headers', 'validate.payload', 'validate.params', 'validate.query']);
         this.settings.handler = handler;
         this.settings = Config.apply('routeConfig', this.settings, routeDisplay);
 
@@ -108,8 +108,8 @@ exports = module.exports = internals.Route = class {
 
             // Assert on config, not on merged settings
 
-            Hoek.assert(!config || !config.payload, 'Cannot set payload settings on HEAD or GET request:', routeDisplay);
-            Hoek.assert(!config || !config.validate || !config.validate.payload, 'Cannot validate HEAD or GET requests:', routeDisplay);
+            Hoek.assert(!config.payload, 'Cannot set payload settings on HEAD or GET request:', routeDisplay);
+            Hoek.assert(!config.validate || !config.validate.payload, 'Cannot validate HEAD or GET requests:', routeDisplay);
 
             validation.payload = null;
         }
@@ -382,11 +382,15 @@ internals.payload = async function (request) {
 
         request._isPayloadPending = !!(payload && payload._readableState);
         request.mime = mime;
-        request.payload = (payload === undefined ? null : payload);
+        request.payload = payload;
     }
     catch (err) {
         if (request._isPayloadPending) {
             await internals.drain(request);
+            request._isPayloadPending = false;
+        }
+        else {
+            request._isPayloadPending = true;
         }
 
         request.mime = err.mime;
