+++ lib/toolkit.js
@@ -59,9 +59,9 @@ exports = module.exports = internals.Manager = class {
                 return;
             }
 
-            if (options.continue === 'null') {
-                response = null;
-            }
+            // 'null'
+
+            response = null;
         }
 
         if (options.auth &&
