+++ lib/transmit.js
@@ -132,6 +132,7 @@ internals.length = function (response) {
     if (typeof length === 'string') {
         length = parseInt(header, 10);
         if (!isFinite(length)) {
+            delete response.headers['content-length'];
             return null;
         }
     }
