+++ test/server.js
@@ -261,6 +261,34 @@ describe('Server', () => {
             expect(res.result).to.equal('testing123');
         });
 
+        it('register a plugin once (plugin options)', async () => {
+
+            let count = 0;
+            const b = function (srv, options) {
+
+                ++count;
+            };
+
+            b.attributes = {
+                name: 'b'
+            };
+
+            const a = async function (srv, options) {
+
+                await srv.register({ register: b, once: true });
+            };
+
+            a.attributes = {
+                name: 'a'
+            };
+
+            const server = new Hapi.Server();
+            await server.register(b);
+            await server.register(a);
+            await server.initialize();
+            expect(count).to.equal(1);
+        });
+
         it('registers plugins and adds options to realm that routes can access', async () => {
 
             const server = new Hapi.Server();
@@ -1387,7 +1415,7 @@ describe('Server', () => {
             const server = new Hapi.Server();
 
             const updates = [];
-            server.events.registerEvent({ name: 'test', channels: ['x', 'y'] });
+            server.event({ name: 'test', channels: ['x', 'y'] });
 
             server.events.on({ name: 'test', channels: 'x' }, (update) => updates.push({ id: 'server', channel: 'x', update }));
